+++ src/util/OrmUtils.ts
@@ -2,6 +2,15 @@ import {ObjectLiteral} from "../common/ObjectLiteral";
 
 export class OrmUtils {
 
+    /**
+     * Chunks array into peaces.
+     */
+    static chunk<T>(array: T[], size: number): T[][] {
+        return Array.from(Array(Math.ceil(array.length / size)), (_, i) => {
+            return array.slice(i * size, i * size + size);
+        });
+    }
+
     static splitClassesAndStrings<T>(clsesAndStrings: T[]|string[]): [T[], string[]] {
         return [
             (clsesAndStrings as T[]).filter(cls => typeof cls !== "string"),